refactor: extracts interface from RowMapper
